docker-requirements.txt
allure-pytest==2.8.36
allure-pytest-bdd==2.8.36
allure-python-commons==2.8.36
pandas==1.2.3
pytest==6.2.2
pytest-bdd==4.0.2
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-rerunfailures==9.1.1
pytest-xdist==2.2.1
requests==2.25.1
selenium==3.141.0
XlsxWriter==1.3.7
zeep==4.0.0
msedge-selenium-tools==3.141.3
selenium-page-factory==2.2
openpyxl==3.0.7
jsonpath==0.82
SQLAlchemy==1.4.12
pymssql==2.2.1
PyMySQL==1.0.2
pyodbc==4.0.30
cx-Oracle==8.1.0
webdriver-manager==3.4.2
Pillow==8.1.2
pytest-assume==2.4.3
lxml==4.6.3
beautifulsoup4==4.10.0
fuzzywuzzy==0.18.0
Appium-Python-Client==1.2.0
rstr==2.2.6
Faker==8.10.3
nltk==3.6.2
joblib==1.0.1
scikit-learn==0.24.2
Jinja2==3.0.1